+++ itsim/node/process_management/thread.py
@@ -1,5 +1,5 @@
 from itsim.simulator import Simulator
-from itsim.node.processes import _Process, _Thread
+from itsim.node.process_management import _Process, _Thread
 
 from typing import Callable, Set
