+++ sphinx/source/index.rst
@@ -14,10 +14,16 @@ ITSim is a discrete event simulation of IT infrastructures.
    networks
    internet
    endpoints
+   functions
 
 The source code is all available on `Github <https://github.com/ElementAI/itsim>`_ and it can be built on your machine by following the :doc:`build_instructions`
 
-In addition, take a gander at the :doc:`overview` to get an understanding of our language design from the top level.
+Background
+----------
+
+Take a gander at the :doc:`overview` to get an understanding of the language design from the top level.
+
+The document :doc:`functions` gives special treatment to the special treatment of Python functions in the language design
 
 How-to
 -------
